Adds custom hook call
