+++ src/components/Connect4/gameContextProvider.jsx
@@ -1,13 +1,13 @@
 import { useContext, useState, useEffect } from "react";
 import { appContext } from "./../../AppContext";
-import { useParams, useLocation } from "react-router-dom";
-import axios from "axios";
+import { useParams } from "react-router-dom";
 import "./../../App.css";
 import "bootstrap/dist/css/bootstrap.min.css";
 import { gameContext } from "./gameContext";
+import usePopulateGameData from "../../hooks/usePopulateGameData";
 
 export const GameContextProvider = (props) => {
-  const { socket, user } = useContext(appContext);
+  const { socket } = useContext(appContext);
   const numRows = 8; // needs to be dynamic
   const numCols = 6; // needs to be dynamic
   const [moves, setMoves] = useState([]);
@@ -23,24 +23,13 @@ export const GameContextProvider = (props) => {
     )
   );
   let { id } = useParams();
-  const location = useLocation();
-  useEffect(() => {
-    const populateGameData = (gameData) => {
-      setGameId(gameData.id);
-      setTurn(gameData.turn === 1 ? "blue" : "red");
-      setFilled(gameData.filled);
-      setAllowed(gameData.allowed);
-    };
-    axios
-      .get(process.env.REACT_APP_SERVER_URL + location.pathname, {
-        headers: {
-          Authorization: "Bearer " + user,
-        },
-      })
-      .then((res) => res.data)
-      .then((gameData) => populateGameData(gameData))
-      .catch((err) => console.error(err.message));
-  }, []);
+  const populateGameData = (gameData) => {
+    setGameId(gameData.id);
+    setTurn(gameData.turn === 1 ? "blue" : "red");
+    setFilled(gameData.filled);
+    setAllowed(gameData.allowed);
+  };
+  usePopulateGameData(populateGameData);
   useEffect(() => {
     function handleConnect4MoveSuccess(data) {
       console.log(`a move was played - ${JSON.stringify(data)}`);
